Clear up useless and circular imports
